refactor: rename sharedcore references to commoncore for consistency
