+++ chatbot/src/main/java/pt/estga/chatbot/models/BotInput.java
@@ -2,7 +2,7 @@ package pt.estga.chatbot.models;
 
 import lombok.Builder;
 import lombok.Data;
-import pt.estga.sharedcore.models.Location;
+import pt.estga.commoncore.models.Location;
 
 import java.util.Locale;
