+++ chatbot/src/main/java/pt/estga/chatbot/whatsapp/WhatsAppMapper.java
@@ -3,7 +3,7 @@ package pt.estga.chatbot.whatsapp;
 import lombok.extern.slf4j.Slf4j;
 import pt.estga.chatbot.models.BotInput;
 import pt.estga.chatbot.models.Platform;
-import pt.estga.sharedcore.models.Location;
+import pt.estga.commoncore.models.Location;
 
 @Slf4j
 public class WhatsAppMapper {
