+++ chatbot/src/test/java/pt/estga/chatbot/services/BotEngineImplTest.java
@@ -17,8 +17,8 @@ import pt.estga.chatbot.context.CoreState;
 import pt.estga.chatbot.models.BotInput;
 import pt.estga.chatbot.models.BotResponse;
 import pt.estga.chatbot.models.Platform;
-import pt.estga.sharedcore.models.AppPrincipal;
-import pt.estga.sharedcore.utils.SecurityUtils;
+import pt.estga.commoncore.models.AppPrincipal;
+import pt.estga.commoncore.utils.SecurityUtils;
 
 import java.util.List;
 import java.util.Optional;
