+++ common-core/src/main/java/pt/estga/commoncore/enums/EntityStatus.java
@@ -1,4 +1,4 @@
-package pt.estga.sharedcore.enums;
+package pt.estga.commoncore.enums;
 
 public enum EntityStatus {
     ACTIVE,
