+++ common-core/src/main/java/pt/estga/commoncore/enums/ValidationState.java
@@ -1,4 +1,4 @@
-package pt.estga.sharedcore.enums;
+package pt.estga.commoncore.enums;
 
 /**
  * Centralized validation state used across multiple modules.
