+++ common-core/src/main/java/pt/estga/commoncore/events/AfterCommitEventPublisher.java
@@ -1,4 +1,4 @@
-package pt.estga.sharedcore.events;
+package pt.estga.commoncore.events;
 
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
