+++ common-core/src/main/java/pt/estga/commoncore/interfaces/AuthenticatedPrincipal.java
@@ -1,4 +1,4 @@
-package pt.estga.sharedcore.interfaces;
+package pt.estga.commoncore.interfaces;
 
 public interface AuthenticatedPrincipal {
     Long getId();
