+++ common-core/src/main/java/pt/estga/commoncore/security/Permissions.java
@@ -1,4 +1,4 @@
-package pt.estga.sharedcore.security;
+package pt.estga.commoncore.security;
 
 public final class Permissions {
