+++ common-core/src/main/java/pt/estga/commoncore/utils/JwtSubExtractor.java
@@ -1,4 +1,4 @@
-package pt.estga.sharedcore.utils;
+package pt.estga.commoncore.utils;
 
 import java.util.Map;
