+++ common-core/src/test/java/pt/estga/commoncore/utils/JwtSubExtractorTest.java
@@ -1,4 +1,4 @@
-package pt.estga.sharedcore.utils;
+package pt.estga.commoncore.utils;
 
 import org.junit.jupiter.api.Test;
