+++ mark/src/main/java/pt/estga/mark/entities/Mark.java
@@ -4,7 +4,7 @@ import jakarta.persistence.*;
 import lombok.*;
 import pt.estga.commoninfra.converters.ValidationStateConverter;
 import pt.estga.commoninfra.entities.BaseEntity;
-import pt.estga.sharedcore.enums.ValidationState;
+import pt.estga.commoncore.enums.ValidationState;
 
 import java.util.List;
