+++ user/src/main/java/pt/estga/user/controllers/AccountController.java
@@ -13,7 +13,7 @@ import org.springframework.http.ResponseEntity;
 import org.springframework.security.access.prepost.PreAuthorize;
 import org.springframework.security.core.annotation.AuthenticationPrincipal;
 import org.springframework.web.bind.annotation.*;
-import pt.estga.sharedcore.interfaces.AuthenticatedPrincipal;
+import pt.estga.commoncore.interfaces.AuthenticatedPrincipal;
 import pt.estga.commonweb.dtos.MessageResponseDto;
 import pt.estga.user.dtos.*;
 import pt.estga.user.services.UserService;
